// src/urn.d.ts
declare const namespace: {
    binding: {
        redirect: string;
        post: string;
        simpleSign: string;
        artifact: string;
        soap: string;
    };
    names: {
        protocol: string;
        assertion: string;
        metadata: string;
        userLogout: string;
        adminLogout: string;
    };
    authnContextClassRef: {
        password: string;
        passwordProtectedTransport: string;
    };
    format: {
        emailAddress: string;
        persistent: string;
        transient: string;
        entity: string;
        unspecified: string;
        kerberos: string;
        windowsDomainQualifiedName: string;
        x509SubjectName: string;
    };
    statusCode: {
        success: string;
        requester: string;
        responder: string;
        versionMismatch: string;
        authFailed: string;
        invalidAttrNameOrValue: string;
        invalidNameIDPolicy: string;
        noAuthnContext: string;
        noAvailableIDP: string;
        noPassive: string;
        noSupportedIDP: string;
        partialLogout: string;
        proxyCountExceeded: string;
        requestDenied: string;
        requestUnsupported: string;
        requestVersionDeprecated: string;
        requestVersionTooHigh: string;
        requestVersionTooLow: string;
        resourceNotRecognized: string;
        tooManyResponses: string;
        unknownAttrProfile: string;
        unknownPrincipal: string;
        unsupportedBinding: string;
    };
};
// ...
declare const messageConfigurations: {
    signingOrder: {
        SIGN_THEN_ENCRYPT: string;
        ENCRYPT_THEN_SIGN: string;
    };
};
declare const algorithms: {
    signature: {
        RSA_SHA1: string;
        RSA_SHA256: string;
        RSA_SHA512: string;
    };
    encryption: {
        data: {
            AES_128: string;
            AES_256: string;
            AES_256_GCM: string;
            TRI_DEC: string;
            AES_128_GCM: string;
        };
        key: {
            RSA_OAEP_MGF1P: string;
            RSA_1_5: string;
        };
    };
    digest: {
        'http://www.w3.org/2000/09/xmldsig#rsa-sha1': string;
        'http://www.w3.org/2001/04/xmldsig-more#rsa-sha256': string;
        'http://www.w3.org/2001/04/xmldsig-more#rsa-sha512': string;
    };
};
export declare enum ParserType {
    SAMLRequest = "SAMLRequest",
    SAMLResponse = "SAMLResponse",
    LogoutRequest = "LogoutRequest",
    LogoutResponse = "LogoutResponse"
}
declare const wording: {
    urlParams: {
        samlRequest: string;
        samlResponse: string;
        logoutRequest: string;
        logoutResponse: string;
        sigAlg: string;
        signature: string;
        relayState: string;
    };
    binding: {
        redirect: string;
        post: string;
        simpleSign: string;
        artifact: string;
        soap: string;
    };
    certUse: {
        signing: string;
        encrypt: string;
    };
    metadata: {
        sp: string;
        idp: string;
    };
};